requirements.txt
awsebcli==3.15.0
blessed==1.15.0
cached-property==1.5.1
cement==2.8.2
chardet==3.0.4
colorama==0.3.9
Django==2.2.4
django-admin-sortable2==0.7.1
django-autocomplete-light==3.3.4
django-bootstrap4==0.0.8
django-filter==2.1.0
django-multiselectfield==0.1.8
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==2.7
jmespath==0.9.4
pathspec==0.5.9
Pillow==6.2.0
pytz==2019.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
websocket-client==0.56.0
schedule
django-ckeditor
pushbullet.py
celery[redis]
django-appconf
ipaddress
slacker
django-debug-toolbar